feat: Congratulatory modal
Implemented a modal that shows up once the player has won the game, it'll show their star rating, the number of moves they used and how long it took them to complete the game.

It'll also prompt them whether they'd like to play again or not, if yes then the game restarts otherwise just close the modal.

Resolves: #14